+++ nextstep-di/src/main/java/nextstep/di/factory/beancreator/BeanCreator.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory;
+package nextstep.di.factory.beancreator;
 
 import java.util.List;
