+++ nextstep-di/src/main/java/nextstep/di/factory/beancreator/ClassBeanCreator.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory;
+package nextstep.di.factory.beancreator;
 
 import com.google.common.collect.Lists;
 import nextstep.annotation.Inject;
