+++ nextstep-di/src/main/java/nextstep/di/factory/beancreator/MethodBeanCreator.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory;
+package nextstep.di.factory.beancreator;
 
 import com.google.common.collect.Lists;
 import nextstep.di.factory.exception.ObjectInstantiationFailException;
